+++ src/main/java/pl/droidsonroids/gif/GifInfoHandle.java
@@ -13,6 +13,8 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.nio.ByteBuffer;
 
+import static pl.droidsonroids.gif.GifOptions.UINT16_MAX;
+
 /**
  * Native library wrapper
  */
@@ -184,8 +186,8 @@ final class GifInfoHandle {
 		return getLoopCount(gifInfoPtr);
 	}
 
-	void setLoopCount(@IntRange(from = 0, to = 0xFFFF) final int loopCount) {
-		if (loopCount < 0 || loopCount > 0xFFFF) {
+	void setLoopCount(@IntRange(from = 0, to = UINT16_MAX) final int loopCount) {
+		if (loopCount < 0 || loopCount > UINT16_MAX) {
 			throw new IllegalArgumentException("Loop count of range <0, 65535>");
 		}
 		synchronized (this) {